pyproject.toml
[tool.poetry]
name = "seminario1-api"
version = "0.1.0"
description = ""
authors = ["Alexander Cabezas <acabezas@inmarket.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
keras = "^2.6.0"
tensorflow-cpu = "^2.6.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
